Move secret code into utils
