+++ sdks/java/core/src/main/java/org/apache/beam/sdk/util/Secret.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.transforms;
+package org.apache.beam.sdk.util;
 
 import java.io.Serializable;
